created smDoc - SnipMate Documentation tool.
smDoc creates HTML and PDF documentation from snipMate files. I created it because I find that I don't really take the time to learn the snippets - having a paper copy means I    can carry it around and learn the snippet commands while on the train.

HTML generation is done using the smdoc.rb tool - invoked using `make doc-html`.

The HTML documentation generation uses a simple JavaDoc-style syntax - for each snippet entry, you can prefix it with a line that starts "## ". In there, put as much description as you want to. You can use inline HTML if you like, but try to keep the descriptions short and snappy. It renders these into a definition list (the dl,dt,dd tags).

As well as this, you need to produce a small file called metadata.yml which contains the name of the programming language or tool you have written the snippets file for, the name of the primary author and a contact e-mail address. metadata.yml may contain other stuff which gets added in the future (excludes, etc.)

PDF generation is currently for OS X only. It uses a RubyGem called wkpdf which uses the WebKit rendering engine on the Mac to render a PDF. This can be invoked using `make doc-pdf-a4` and `make doc-pdf-letter`. There is a Linux alternative to wkpdf called wkhtmltopdf at <http://code.google.com/p/wkhtmltopdf/> but it looks quite complicated to get setup.

I chose to use the WebKit-based renderer as an alternative to HTMLDOC because HTMLDOC doesn't render CSS, while the WebKit-based renderers actually render CSS and JavaScript stuff properly.